Day 11 - Part 1 WIP - Extract gaps - split structs into their mods
